feat(privacy-policy): move dispatch to redux
