+++ src/PrivacyPolicy/PrivacyPolicy.less
@@ -1,6 +1,8 @@
 .privacy-policy {
   padding: 40px 40px;
 
+  // not following BEM here since the HTML from the privacy is dynamically
+  // generated and we can't add class names to it
   ul {
     margin-left: 20px;
     list-style-type: disc;
