+++ src/PrivacyPolicy/ReduxPrivacyPolicy.js
@@ -1,4 +1,5 @@
 import { connect } from 'react-redux';
+import showdown from 'showdown';
 import PrivacyPolicy from './PrivacyPolicy';
 
 const mapStateToProps = state => ({
@@ -7,7 +8,20 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  dispatch,
+  loadPrivacyPolicy: () => {
+    fetch('/src/privacy_policy.md').then(
+      response => response.text().then(
+        (text) => {
+          const converter = new showdown.Converter();
+          const html = converter.makeHtml(text);
+          dispatch({
+            type: 'LOAD_PRIVACY_POLICY',
+            value: html,
+          });
+        }),
+      _ => '', // eslint-disable-line no-unused-vars
+    );
+  },
 });
 
 const ReduxPrivacyPolicy = connect(mapStateToProps, mapDispatchToProps)(PrivacyPolicy);
